REWRITE LOOP IN MAIN AND CLEANUP
Rewrite loop in main to detect preprocessor directives and identifiers
in a less error prone way by using one variable with a state (and
defines for what the state means) instead of many different ints that
represent boolean on and offs. This is way less complex because less
variables and also no possiblity for it to be in multiple states at
once. For instance lets say we have variable identifier and preprocessor
directive as ints but represent booleans values. Both can be true at the
same time meaning more complex code to ensure they are not both true at
the same time.

Also:
* removed other unneccessary variables
* set up the basic and hopefully final layout for the main loop
* simpler use no word to store identifier or define statement and add to
  buffer, just keep in buffer and record where it is

Note: I should have probably done this in two or three seperate
commits. Remember this for next time!